+++ app/Http/Controllers/SizeController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 use App\Models\Size;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
-
+use Illuminate\Support\Facades\Gate;
 
 class SizeController extends Controller
 
@@ -12,6 +12,9 @@ class SizeController extends Controller
 
     public function index()
     {
+        if (!Gate::allows('hasRole', ['Admin','Manager'])) {
+            abort(403, 'Unauthorized');
+        }
         $allsize =Size::orderBy('created_at', 'desc')->get();//
 
         return Inertia::render('Size/Index', [
@@ -22,6 +25,10 @@ class SizeController extends Controller
 
     public function store(Request $request)
     {
+
+        if (!Gate::allows('hasRole', ['Admin'])) {
+            abort(403, 'Unauthorized');
+        }
         $validated = $request->validate([
             'name' => 'required|string|max:255',
 
@@ -39,6 +46,10 @@ class SizeController extends Controller
 
     public function update(Request $request, Size $Size)
     {
+
+        if (!Gate::allows('hasRole', ['Admin'])) {
+            abort(403, 'Unauthorized');
+        }
         $validated = $request->validate([
             'name' => 'nullable|string|max:255',
         ]);
@@ -54,6 +65,9 @@ class SizeController extends Controller
     public function destroy(Size $size)
     {
 
+        if (!Gate::allows('hasRole', ['Admin'])) {
+            abort(403, 'Unauthorized');
+        }
         $size->delete();
         return redirect()->route('sizes.index')->banner('Size Deleted successfully.');
     }
